make a float to ensure correct ops when used as a divisor

+++ trainGa.py
@@ -207,7 +207,7 @@ if __name__ == '__main__':
     parser.add_argument('--logging_start', type=int, default=1, help="First epoch to be logged")
     parser.add_argument('--max_gpus', type=int, default=2, help="Maximal number of GPUs of the local machine to use.")
     parser.add_argument('--loader_workers', type=int, default=0, help="Number of subprocesses to use for data loading.")
-    parser.add_argument("--accumulation_size", type=int, default=1, help="How many batches to perform gradient accumulation over.")
+    parser.add_argument("--accumulation_size", type=float, default=1.0, help="How many batches to perform gradient accumulation over.")
     parser.add_argument('--fine_tuning', action='store_true', help="Fine tune checkpoint to possibly unseen language.")
     parser.add_argument('--log_high_loss', action='store_true', help="Log batch details for high loss values.")
     args = parser.parse_args()
